* master:
  core: check for IP in no_naptr_srv_sip_resolvehost()
  core: try all srv protocols when not already set
  registrar(k): restore the location contact xavp upon lookup
  usrloc(k): option to store xavp per contact
  core: extended xavp api
  registrar(k): use only pre-existing branches in lookup_branches()
  p_usrloc: clean memory in case of malloc error
  Revert "Changing to kamcmd"
  kamcmd changes
  Changing to kamcmd
  kamcmd Updating docs
  Change "sercmd" to "kamcmd" in module documentation files - README
  kamctl Fix typo that causes the "trusted" help to be printed twice and "address" zero times
  kamctl Change "sercmd" to "kamcmd" in help texts
  Fixing compiler warnings

+++ modules_k/pike/pike_top.c
@@ -45,7 +45,7 @@ static char *print_addr(unsigned char *ip, int iplen)
 	return pike_top_print_addr(ip, iplen, buff, sizeof(buff));
 }
 
-int pike_top_add_entry( unsigned char *ip_addr, int addr_len, unsigned int leaf_hits[2], unsigned int hits[2], unsigned int expires, node_status_t status )
+int pike_top_add_entry( unsigned char *ip_addr, int addr_len, unsigned short leaf_hits[2], unsigned short hits[2], unsigned int expires, node_status_t status )
 {
 	struct TopListItem_t *new_item = (struct TopListItem_t *)malloc(sizeof(struct TopListItem_t));
